+++ ricaun.NUnit.Tests/ReferenceLoaderUtils_Tests.cs
@@ -4,6 +4,8 @@ using System;
 using System.Linq;
 using System.Reflection;
 
+[assembly: AssemblyMetadata("Tests", "Tests")]
+
 namespace ricaun.NUnit.Tests
 {
     public class ReferenceLoaderUtils_Tests
@@ -44,5 +46,23 @@ namespace ricaun.NUnit.Tests
             var endWith = AppDomain.CurrentDomain.GetAssemblies().Length;
             Assert.Zero(endWith - startWith);
         }
+
+        [Test]
+        public void ReferenceLoaderUtils_Tests_GetMetadataAttributes_Assemblies()
+        {
+            var assemblyFile = Assembly.GetExecutingAssembly().Location;
+            var startWith = AppDomain.CurrentDomain.GetAssemblies().Length;
+
+            var assemblyMetadataAttributes = ReferenceLoaderUtils.GetAssemblyMetadataAttributes(assemblyFile);
+            foreach (var assemblyMetadataAttribute in assemblyMetadataAttributes)
+            {
+                Console.WriteLine($"{assemblyMetadataAttribute.Key}: {assemblyMetadataAttribute.Value}");
+            }
+            Assert.IsNotNull(assemblyMetadataAttributes.FirstOrDefault(e => e.Key.Equals("Tests")));
+
+            var endWith = AppDomain.CurrentDomain.GetAssemblies().Length;
+            Assert.Zero(endWith - startWith);
+        }
     }
+
 }