+++ ricaun.NUnit/Extensions/ReferenceLoaderUtils.cs
@@ -1,4 +1,5 @@
 ﻿using System.IO;
+using System.Linq;
 using System.Reflection;
 
 namespace ricaun.NUnit.Extensions
@@ -30,6 +31,31 @@ namespace ricaun.NUnit.Extensions
         {
             var assembly = Assembly.Load(File.ReadAllBytes(assemblyPath));
             return assembly.GetReferencedAssemblies();
+        }
+
+#if NETFRAMEWORK
+#elif NET
+#else
+        /// <summary>
+        /// GetAssemblyMetadataAttributes by Load Assembly with <see cref="File.ReadAllBytes"/>
+        /// </summary>
+        /// <param name="assemblyPath"></param>
+        /// <returns></returns>
+        public static AssemblyMetadataAttribute[] GetAssemblyMetadataAttributes(string assemblyPath)
+        {
+            return GetAssemblyMetadataAttributesDefault(assemblyPath);
+        }
+#endif
+
+        /// <summary>
+        /// GetAssemblyMetadataAttributesDefault by Load Assembly with <see cref="File.ReadAllBytes"/>
+        /// </summary>
+        /// <param name="assemblyPath"></param>
+        /// <returns></returns>
+        internal static AssemblyMetadataAttribute[] GetAssemblyMetadataAttributesDefault(string assemblyPath)
+        {
+            var assembly = Assembly.Load(File.ReadAllBytes(assemblyPath));
+            return assembly.GetCustomAttributes<AssemblyMetadataAttribute>().ToArray();
         }
     }
 }