+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.4.1] / 2024-12-07
+### Updates
+- Add `TestEngine.GetAssemblyMetadataAttributes` to get `AssemblyMetadataAttribute` from `assemblyFile`.
+
 ## [1.4.0] / 2024-09-11
 ### Features
 - Support `TestCaseSourceAttribute` with `IEnumerable` and `object[]` parameters. (Fix: #27)
@@ -236,7 +240,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [1.0.0] / 2022-08-18
 ### Features
-- Test Assembly with diferent NUnit Version with LoadFile
+- Test Assembly with different NUnit Version with LoadFile
 - Test Methods with Attribute Name
 - Test if Assembly contain the same NUnit Version
 - Test Method with custom parameters
@@ -244,6 +248,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Test with Console Output
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.4.1]: ../../compare/1.4.0...1.4.1
 [1.4.0]: ../../compare/1.3.1...1.4.0
 [1.3.1]: ../../compare/1.3.0...1.3.1
 [1.3.0]: ../../compare/1.2.9...1.3.0